requirements.txt
requests
python-dotenv
pandas
tqdm
google-generativeai
sqlalchemy
psycopg2
numpy
